==> 01_find_coffeeproducts.ipynb <==
import os
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from utils import add_delimiters

products = []

years = ["2021", "2022"]
months = ["01jan", "02feb", "03mar", "04apr", "05may", "06jun", "07jul", "08aug", "09sep", "10oct", "11nov", "12dec"]

for year in years:
    for month in months:
        data_dir = os.path.join('raw_data', year, month)
        data_names = os.listdir(data_dir)
        for day_file in data_names:
            df = pd.read_csv(add_delimiters(os.path.join(data_dir, day_file), delimiter=';'), sep=';', encoding="utf-8")

            index = df.index[df.eq("Produkte").any(1)][0]

            todrop_rows = list(range(index + 1))
            df = df.drop(todrop_rows)

            todrop_columns = list(df.columns[4:])
            df = df.drop(columns = todrop_columns)
            df = df.drop(columns = df.columns[1])
            df = df.dropna()

            df = df.reset_index(drop = True)

            df.columns = ['Product', 'Count', 'Sales']

            for product in df['Product']:
                if product in products:
                    continue
                else:
                    products.append(product)

products
  

# Save Productlist to File

file = open('products.txt','w', encoding="utf-8")
for product in products:
	file.write(product+"\n")
file.close()
